Adding in url query parsing
Now on game start, the URL will read the query string for game data and load that chapter and scene
Moving between scenes and chapter progression update this url.

Modifying scene resolvers to store the actual scene now, and each scene is created with a default spawnpoint based on the chapter.